refactor: profile page separated in components
